refactor: reorganized index config edit module.
replace config json editor with monaco-editor.
prepare to enhance visualization support gh-3.
