+++ src/components/Settings/baseInfo.css
