+++ src/components/Settings/index.ts
@@ -1,0 +1,7 @@
+import { Index } from 'meilisearch';
+
+export type IndexSettingComponentProps = {
+  className?: string;
+  host?: string;
+  client: Index;
+};
